Revert "moved all figures to array, added figure, that spins vertically, some other minor changes"
This reverts commit 86529106ed5c638dfe2185139c2f78931240c2ff.

+++ Allegro_Spin/PolygonFactory.hpp
@@ -3,6 +3,6 @@
 #include "Poly2d.hpp"
 class PolygonFactory {
 public:
-    static Poly2d NewPolygon4Points(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4, float t0, ALLEGRO_COLOR col);
-    static Poly2d NewPolygon(std::vector<float> poly0, float t0, ALLEGRO_COLOR col);
+    static Poly2d NewPolygon4Points(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4, int t0, ALLEGRO_COLOR col);
+    static Poly2d NewPolygon(std::vector<float> poly0, int t0, ALLEGRO_COLOR col);
 };
